+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add rich tooltips to Active Effects
 - Add scrolling text for updates to curse dice, armor, and injuries
 - Add injury dice and enhancement for Near Death level
+- Add integrity rolls to Accursed
+- Increase the Defense and Integrity values of Accursed when buying Dodge/Resist tracks for their respective rolls
 
 ### Changed